add tests for same name/same build
